Update on items
Adding some things for feedback, we can discuss their inclusion.

For the Hub and Spoke / VWAN - I phrased it as hub/spoke as the primary topology.  This way if there is a question about vwan we can say that vwan uses the same topology mechanics, and let it proceed from there.

Added language about the NSGs not having the explicit allows.