chore: Update ALZ-Policies-FAQ.md with deployment instructions
The commit updates the ALZ-Policies-FAQ.md file to include additional instructions for deploying specific policies in Azure Policy. It advises users to use AzAdvertizer to lookup the policy and copy a clean version of the policy definition, removing the extra leading `[` character. This change aims to provide clearer guidance for deploying policies from the repository.

Co-authored-by: Jack Tracey <41163455+jtracey93@users.noreply.github.com>